+++ frontend/components/studyDetail/StudyProgressBar.vue
